update: added parallel & intercept routing
