+++ app/@modal/(.)prompts/[id]/page.tsx
@@ -1,0 +1,5 @@
+const PromptModal = ({ params }: { params: { id: number } }) => {
+  return <div>prompt page {params.id}</div>;
+};
+
+export default PromptModal;
