+++ app/@modal/default.tsx
@@ -1,0 +1,4 @@
+const Default = () => {
+  return null;
+};
+export default Default;
