+++ app/layout.tsx
@@ -1,9 +1,6 @@
 import './globals.css';
-import { Inter } from 'next/font/google';
 import NavBar from './components/NavBar';
 import React from 'react';
-
-const inter = Inter({ subsets: ['latin'] });
 
 export const metadata = {
   title: 'Find your prompt',
@@ -12,17 +9,17 @@ export const metadata = {
 
 export default function RootLayout({
   children,
-  feed,
+  modal,
 }: {
   children: React.ReactNode;
-  feed: React.ReactNode;
+  modal: React.ReactNode;
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>
+      <body>
         <NavBar />
         {children}
-        {feed}
+        {modal}
       </body>
     </html>
   );
