+++ app/loading.tsx
@@ -1,0 +1,4 @@
+export default function RootLoading() {
+  // You can add any UI inside Loading, including a Skeleton.
+  return <p>Loading...</p>;
+}
