+++ app/page.tsx
@@ -1,4 +1,4 @@
-// import Feed from './components/Feed';
+import Feed from './components/Feed';
 import Hero from './components/Hero';
 import Search from './components/Search';
 import Card from './ui/Card';
@@ -8,7 +8,7 @@ export default function Home() {
     <Card>
       <Hero />
       <Search />
-      {/* <Feed /> */}
+      <Feed />
     </Card>
   );
 }
